modifying services and added photoservice logics

+++ src/main/java/com/beeswork/balanceaccountservice/service/swipe/SwipeServiceImpl.java
@@ -94,7 +94,7 @@ public class SwipeServiceImpl extends BaseServiceImpl implements SwipeService {
 
     @Override
     @Transactional
-    public ClickDTO click(Long swipeId, String accountId, String identityToken, String swipedId, Map<Long, Boolean> answers) {
+    public ClickDTO click(Long swipeId, String accountId, String identityToken, String swipedId, Map<Integer, Boolean> answers) {
 
         UUID swiperUUId = UUID.fromString(accountId);
         UUID swipedUUId = UUID.fromString(swipedId);
